pyproject.toml
[tool.poetry]
name = "create_slack_bday_chat"
version = "0.0.1"
description = "Create birthday channel in slack using this bot. That means that you can create channel without one person."
authors = ["Nikita Churikov <nikita@chur.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
loguru = "^0.4.1"
environs = "^7.1.0"
slackclient = "^2.5.0"
slackeventsapi = "^2.1.0"
certifi = "^2019.11.28"
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.black]
line-length = 100
target_version = ["py37"]

[build-system]
requires = ["poetry==1.*"]
build-backend = "poetry.masonry.api"
